backend/requirements.txt
flask
flask-cors
torch
torchvision
transformers
huggingface-hub
pillow
pymongo
gunicorn
